Add validate subcommand (#66)
Add a validate subcommand and pre-commit hook to validate the configuration. This should help catch the occasional bug or mishap i create. This is mostly for #3 so that if i do add bad config the site will not break. Closes #63.

I've also cleaned up the `argparse` set up because i finally hit 3 things and it was annoying me.