Improve useability of prv/next buttons
The tiny little arrows sucked for pressing on mobile and were annoying
to hit with the mouse. Rework the doc structure and CSS to make the
targets the entire html element instead of just the text.

I also improved the template so we don't generate any html in the code
and its all in template now.