Add a better workflow for images (#102)
Convert to using uv / actually set this up as a tool.

### Workflow Updates:
*
[`.github/workflows/pages.yml`](diffhunk://#diff-267452bb53d388a7b14551d9fc53a797edc124dc9085ecc79f5339277a1e53b5L31-R46):
Updated the workflow to include `uv` installation and changed the Python
setup to use the version specified in `pyproject.toml`.
*
[`.github/workflows/pre-commit.yml`](diffhunk://#diff-ac5eead3f3ce4863c524fff031a87b7aecccb4a0493df087a4e1c704a1505036L10-R15):
Updated the workflow to use the latest versions of `actions/checkout`
and `actions/setup-python`, and changed the Python version to be
specified in `pyproject.toml`.

### Dependency Management:
*
[`.pre-commit-config.yaml`](diffhunk://#diff-63a9c44a44acf85fea213a857769990937107cf072831e1a26808cfde9d096b9L3-R42):
Switched to use `ruff` and `ruff-format` given the change to uv.
*
[`.python-version`](diffhunk://#diff-652bad7c1002badcf13f8de3a5096c2724ce54ae36cd68a21ad280547dff8e87R1):
Updated the Python version to 3.13.
*
[`pyproject.toml`](diffhunk://#diff-50c86b7ed8ac2cf95bd48334961bf0530cdc77b5a56f852c5c61b89d735fd711R1-R13):
Added project metadata and dependencies

### New Functionality:
*
[`src/dailyphoto/__main__.py`](diffhunk://#diff-b5f5879b27ab3565e5d0573d946bf7f09b9c511d1a06fc9d1fd91fdbb750dd59R1-R4):
Added an entry point for the `dailyphoto` CLI.
*
[`src/dailyphoto/cli.py`](diffhunk://#diff-6034dd961c36c2272b69264c7e69b706e89cceea3087c62a43b6db09ebf40392R1-R93):
Implemented a CLI for generating the photo site, adding new images,
validating configurations, updating metadata, and queuing images.
*
[`src/dailyphoto/generate.py`](diffhunk://#diff-352bbaa28edacf011b6f1f5b16a33f76d201a543b4ef49874575bcec2c4f4296R1-R248):
Split up original daily.photo script into several files
*
[`src/dailyphoto/metadata.py`](diffhunk://#diff-193cbc2b88813539896c4576194d7eb42580b29435f725a7e6543dd9ae4b262bL1-R7):
Renamed from `add-new-metadata.py` and updated to integrate with the new
CLI and `kitty` for editing metadata.