Fix a bunch of bugs + add new (#103)
* Updated `src/dailyphoto/cli.py` to read configuration once and pass it
as an argument to functions instead of reading it multiple times.
* Removed redundant code and improved readability in
`src/dailyphoto/generate.py` by using consistent variable names and
removing unnecessary parameters.
* Moved `METADATA_TEMPLATE` to `src/dailyphoto/config.py` and updated
references in `src/dailyphoto/metadata.py` and
`src/dailyphoto/queued.py`.
* Enhanced `update` function in `src/dailyphoto/metadata.py` to check
for missing fields and prompt for editing if any are found.
* `new_image` was never passed a `None` param in `src/dailyphoto/new.py`
so changed it to respect the `""`.

These changes collectively improve the maintainability and functionality
of the codebase.