fix exif date (#127)
I noticed that "Date" is actually when the image was edited / created
from lightroom / etc so this was often the wrong date. The
"DateTimeOriginal" field appears to be what I actually want. This is
silly.

I've fixed up existing files as well.

Also, queueing images no longer creates the relevant json since we will
just create it when the metadata is created. I've updated new to work
with this.

Added today's image